SiteUser 엔티티 추가
시큐리티에 기본적으로 User 클래스가 존재하기 때문에 헷갈리지 않도록

+++ src/main/java/com/ll/sbb/answer/AnswerController.java
@@ -22,7 +22,7 @@ public class AnswerController {
 
 
     @PostMapping("/create/{id}")
-    public String createAnswer(Model model, @PathVariable int id,
+    public String createAnswer(Model model, @PathVariable Long id,
                                @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = questionService.getQuestion(id);
